chore: 타입 모듈화 및 이름 변경
